prediction_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
mlflow==2.10.0
xgboost==2.0.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
httpx==0.27.2
pytest==8.3.3
